реализован сервис не хватает изменений в клиенте статистики и реквестов

+++ main-service/src/main/java/ru/practicum/event/model/dto/NewEventDto.java
@@ -14,7 +14,7 @@ public class NewEventDto {
     @Size(min = 20, max = 2000)
     private String annotation;
     @NotNull
-    private int category;
+    private long category;
     @NotNull
     @Size(min = 20, max = 7000)
     private String description;
